renamed use-pathogen-map-mouse-events to use-pathogen-map-mouse
